Further work on keystore management.  Added new database table keystore_file and corresponding entity KeystoreFile.  Private key dialog now goes through SsgKeyStoreManager interface to get the available private keys to display.
git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@16177 219f1c6e-4a0e-43c8-96f2-6e749be13fd4

+++ src/com/l7tech/server/security/keystore/SsgKeyFinder.java
@@ -15,8 +15,8 @@ public interface SsgKeyFinder {
         PKCS11_HARDWARE
     }
 
-    /** @return ID of this key store.  Only unique on a particular SSG node. */
-    int getId();
+    /** @return ID of this key store.  Only guaranteed unique on a particular SSG node. */
+    long getId();
 
     /** @return the display name of this key store.  Not necessarily unique.  Never null. */
     String getName();
